Don't resend ACKs
On a read request, when a data packet is received, it is
acknowledged by the client. Once the server receives this ACK
it will send the next data packet.
It the client fails to receive this next data packet it shouldn't
re-ACK, as the server will consider this a "WRONG-BLOCK-NUMBER"
ACK and error out. Or at least this what the JAM/LWIP TFTP
server does.

This commit stops re-ACKing on timeout. If a timeout occurs,
the client should patiently wait for the server to resend the
data packet. It doesn't need to do anything.